System status refactoring.

+++ protected/commands/PackageCommand.php
@@ -58,11 +58,28 @@ class PackageCommand extends ConsoleCommand {
 
         if (flock($fp, LOCK_EX | LOCK_NB)) {
             for ($i = 0; $i < 10; $i++) {
-                try {
-                    $this->_installPackages();
-                    $this->_deletePackages();
-                } catch (Exception $e) {
-                    Yii::log($e->getMessage(), CLogger::LEVEL_ERROR);
+                $this->_system->refresh();
+
+                if ($this->_system->status == System::STATUS_PACKAGE_MANAGER) {
+                    try {
+                        $this->_installPackages();
+                        $this->_deletePackages();
+                    } catch (Exception $e) {
+                        Yii::log($e->getMessage(), CLogger::LEVEL_ERROR);
+                    }
+
+                    $criteria = new CDbCriteria();
+                    $criteria->addInCondition("status", array(Package::STATUS_INSTALL, Package::STATUS_DELETE));
+                    $packages = Package::model()->findAll($criteria);
+
+                    if (count($packages) == 0) {
+                        try {
+                            SystemManager::updateStatus(System::STATUS_IDLE);
+                        } catch (Exception $e) {
+                            // swallow exceptions
+                        }
+
+                    }
                 }
 
                 sleep(5);
